ml_service/pyproject.toml
[project]
name = "ml-service"
version = "0.1.0"
description = "PyTorch ML microservice for window control predictions"
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt>=1.6.1",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
